refactor(utils): share color format utilities
