+++ .changeset/shared-color-format-utils.md
@@ -1,0 +1,6 @@
+---
+'@lapidist/design-lint': patch
+---
+
+refactor color format utilities into shared module
+
